app: move some config functions from admin.go to config.go
