+++ app/admin.go
@@ -12,12 +12,10 @@ import (
 
 	"net/http"
 
-	"github.com/mattermost/mattermost-server/config"
 	"github.com/mattermost/mattermost-server/mlog"
 	"github.com/mattermost/mattermost-server/model"
 	"github.com/mattermost/mattermost-server/services/mailservice"
 	"github.com/mattermost/mattermost-server/utils"
-	"github.com/pkg/errors"
 )
 
 func (a *App) GetLogs(page, perPage int) ([]string, *model.AppError) {
@@ -148,43 +146,6 @@ func (a *App) InvalidateAllCachesSkipSend() {
 	a.Srv.Store.FileInfo().ClearCaches()
 	a.Srv.Store.Webhook().ClearCaches()
 	a.LoadLicense()
-}
-
-func (a *App) GetSanitizedConfig() *model.Config {
-	cfg := a.Config().Clone()
-	cfg.Sanitize()
-
-	return cfg
-}
-
-func (a *App) GetEnvironmentConfig() map[string]interface{} {
-	return a.EnvironmentConfig()
-}
-
-func (a *App) SaveConfig(newCfg *model.Config, sendConfigChangeClusterMessage bool) *model.AppError {
-	oldCfg, err := a.Srv.configStore.Set(newCfg)
-	if errors.Cause(err) == config.ErrReadOnlyConfiguration {
-		return model.NewAppError("saveConfig", "ent.cluster.save_config.error", nil, err.Error(), http.StatusForbidden)
-	} else if err != nil {
-		return model.NewAppError("saveConfig", "app.save_config.app_error", nil, err.Error(), http.StatusInternalServerError)
-	}
-
-	if a.Metrics != nil {
-		if *a.Config().MetricsSettings.Enable {
-			a.Metrics.StartServer()
-		} else {
-			a.Metrics.StopServer()
-		}
-	}
-
-	if a.Cluster != nil {
-		err := a.Cluster.ConfigChanged(oldCfg, newCfg, sendConfigChangeClusterMessage)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
 
 func (a *App) RecycleDatabaseConnection() {
